+++ Sources/SwiftiomaticKit/Syntax/Rules/BlankLines/BlankLinesBetweenScopes.swift
@@ -41,6 +41,7 @@ final class BlankLinesBetweenScopes: RewriteSyntaxRule<BasicRuleValue> {
         var items = original
         var modified = false
         let braceIsBlank = context.configuration[ClosingBraceAsBlankLine.self]
+        let commentIsBlank = context.configuration[CommentAsBlankLine.self]
 
         for i in 0..<(original.count - 1) {
             guard case .decl(let decl) = original[i].item,
@@ -49,6 +50,7 @@ final class BlankLinesBetweenScopes: RewriteSyntaxRule<BasicRuleValue> {
             let nextIndex = i + 1
             guard !original[nextIndex].leadingTrivia.hasBlankLine else { continue }
             if braceIsBlank { continue }
+            if commentIsBlank, original[nextIndex].leadingTrivia.startsWithComment { continue }
 
             diagnose(.insertBlankLineAfterScope, on: original[nextIndex].item)
             var next = original[nextIndex]
@@ -73,12 +75,14 @@ final class BlankLinesBetweenScopes: RewriteSyntaxRule<BasicRuleValue> {
         var modified = false
 
         let braceIsBlank = context.configuration[ClosingBraceAsBlankLine.self]
+        let commentIsBlank = context.configuration[CommentAsBlankLine.self]
 
         for i in 0..<(original.count - 1) {
             guard hasDeclMultiLineBody(original[i].decl) else { continue }
             let nextIndex = i + 1
             guard !original[nextIndex].leadingTrivia.hasBlankLine else { continue }
             if braceIsBlank { continue }
+            if commentIsBlank, original[nextIndex].leadingTrivia.startsWithComment { continue }
 
             diagnose(.insertBlankLineAfterScope, on: diagTargets[nextIndex].decl)
             var next = original[nextIndex]